Changed Docker Builds For Speed/Space Reasons
While I enjoyed how I was able to automate everything about the build proccess inside the Dockerfiles of this project the fact of the matter was that it created a lot of space with <none> images. To fix this <none> images are now removed by the build scripts. I also removed the build proccess of the services outside of docker because it results in much faster build times over all.